requirements.txt
Flask==1.1.2
Flask-Migrate==2.7.0
flask-redis==0.4.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1